Heart Disease Prediction
In this project we want to be able to predict whether a patient should be diagnosed with heart disease through the use of various classification machine learning models  such as LogisticRegression, Support Vector Machine, RandomForestClassifier etc. to see which yields the greatest accuracy. By this, we are able to determine which features are most important to diagnosing heart disease.